refactor: rename DBManager to db_manager for consistency with Dart naming conventions
